# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blocker X - Discord Bot Hosting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with Groq AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIChatResponse"
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /admin/users/{userId}/ban:
    patch:
      operationId: banUser
      tags: [admin]
      summary: Ban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/unban:
    patch:
      operationId: unbanUser
      tags: [admin]
      summary: Unban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/delete:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user account
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/upgrade:
    patch:
      operationId: upgradeUser
      tags: [admin]
      summary: Upgrade user to premium
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/downgrade:
    patch:
      operationId: downgradeUser
      tags: [admin]
      summary: Downgrade user to free
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Downgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/invites:
    get:
      operationId: listInviteCodes
      tags: [admin]
      summary: List all invitation codes
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Invite codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InviteCode"

    post:
      operationId: createInviteCode
      tags: [admin]
      summary: Create an invitation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteCodeCreate"
      responses:
        "201":
          description: Invite code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCode"
  /admin/invites/{inviteId}:
    delete:
      operationId: deleteInviteCode
      tags: [admin]
      summary: Delete an invitation code
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/invites/{inviteId}/toggle:
    patch:
      operationId: toggleInviteCode
      tags: [admin]
      summary: Enable or disable an invitation code
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCode"
  /admin/broadcast:
    post:
      operationId: broadcastAnnouncement
      tags: [admin]
      summary: Broadcast an announcement to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "200":
          description: Announcement sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/logs:
    get:
      operationId: getSystemLogs
      tags: [admin]
      summary: Get system-wide audit logs
      parameters:
        - name: logLimit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: System logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: string
        discordId:
          type: string
        username:
          type: string
        discriminator:
          type: string
        avatar:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, premium]
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        hasInvite:
          type: boolean
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
      required:
        - id
        - discordId
        - username
        - plan
        - isAdmin
        - isBanned
        - hasInvite
        - createdAt
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required:
        - users
        - total
...
    InviteCode:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        maxUses:
          type: ["integer", "null"]
        usesCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdBy:
          type: string
        createdAt:
          type: string
      required:
        - id
        - code
        - usesCount
        - isActive
        - createdBy
        - createdAt
    InviteCodeCreate:
      type: object
      properties:
        maxUses:
          type: ["integer", "null"]
        expiresAt:
          type: ["string", "null"]
        customCode:
          type: ["string", "null"]
      required: []
...
    AIChatInput:
      type: object
      properties:
        message:
          type: string
        botId:
          type: ["string", "null"]
        language:
          type: string
          enum: [python, javascript]
          default: python
        context:
          type: ["string", "null"]
      required:
        - message
    AIChatResponse:
      type: object
      properties:
        response:
          type: string
        tokensUsed:
          type: integer
        usageCount:
          type: integer
        usageLimit:
          type: ["integer", "null"]
      required:
        - response
        - tokensUsed
        - usageCount
...
    AnnouncementInput:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, success, warning, error, announcement]
      required:
        - title
        - message
    AuditLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: ["string", "null"]
        action:
          type: string
        target:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        ip:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - action
        - createdAt