# lib/api-zod/src/generated/types/aIChatInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Blocker X - Discord Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AIChatInputLanguage } from './aIChatInputLanguage';

export interface AIChatInput {
  message: string;
  /** @nullable */
  botId?: string | null;
  language?: AIChatInputLanguage;
  /** @nullable */
  context?: string | null;
}