# lib/api-zod/src/generated/types/inviteCodeCreate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Blocker X - Discord Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface InviteCodeCreate {
  /** @nullable */
  maxUses?: number | null;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  customCode?: string | null;
}